feature/menu
cambio el nombre MenuController por MenuCtr para que quede igual que el resto de controladores
