+++ controladores/gestionPantallasControlador.php
@@ -27,7 +27,7 @@ class GestionPantallasControlador
             case 'presupuestos':
                 // verifico que el tipo de usuario no tiene acceso al modulo y si no lo tiene lo redirijo al menu
                 if (strtoupper($tipoUsuario) != "REPARADOR") {
-                    include_once('./controladores/PresupuestoCtr.php');
+                    include_once ('./controladores/PresupuestoCtr.php');
                     $indexPage = new PresupuestoCtr();
                 } else {
                     $this->redireccionar('menu');
@@ -36,7 +36,7 @@ class GestionPantallasControlador
             case 'reparacion':
                 // verifico que el tipo de usuario no tiene acceso al modulo y si no lo tiene lo redirijo al menu
                 if (strtoupper($tipoUsuario) != "VENDEDOR") {
-                    include_once('controladores/ReparacionControlador.php');
+                    include_once ('controladores/ReparacionControlador.php');
                 } else {
                     $this->redireccionar('menu');
                 }
@@ -44,7 +44,7 @@ class GestionPantallasControlador
             case 'clientes':
                 // verifico que el tipo de usuario no tiene acceso al modulo y si no lo tiene lo redirijo al menu
                 if (strtoupper($tipoUsuario) != "REPARADOR") {
-                    include_once('controladores/ClienteCtr.php');
+                    include_once ('controladores/ClienteCtr.php');
                     $indexPage = new ClienteCtr();
                 } else {
                     $this->redireccionar('menu');
@@ -53,7 +53,7 @@ class GestionPantallasControlador
             case 'proveedores':
                 // verifico que el tipo de usuario no tiene acceso al modulo y si no lo tiene lo redirijo al menu
                 if (strtoupper($tipoUsuario) == "ADMINISTRADOR" || strtoupper($tipoUsuario) == "ADMINISTRADOR BASE") {
-                    include_once('controladores/ProveedorCtr.php');
+                    include_once ('controladores/ProveedorCtr.php');
                     $indexPage = new ProveedorCtr();
                 } else {
                     $this->redireccionar('menu');
@@ -62,7 +62,7 @@ class GestionPantallasControlador
             case 'pedidos':
                 // verifico que el tipo de usuario no tiene acceso al modulo y si no lo tiene lo redirijo al menu
                 if (strtoupper($tipoUsuario) != "REPARADOR") {
-                    include_once('controladores/PedidoCompraControlador.php');
+                    include_once ('controladores/PedidoCompraControlador.php');
                 } else {
                     $this->redireccionar('menu');
                 }
@@ -81,8 +81,8 @@ class GestionPantallasControlador
                 $indexPage = new ProductoCtr();
                 break;
             case 'menu':
-                include_once './controladores/MenuControlador.php';
-                $indexPage = new MenuController();
+                include_once './controladores/MenuCtr.php';
+                $indexPage = new MenuCtr();
                 break;
             default:
                 include_once './vistas/inicio/index.php';
@@ -98,11 +98,11 @@ class GestionPantallasControlador
                     $indexPage->getPantallaDelete();
                     break;
                 case 'create':
-                        $indexPage ->getPantallaCreate();
-                        break;
+                    $indexPage->getPantallaCreate();
+                    break;
                 case 'annul':
-                        $indexPage ->getPantallaAnnul();
-                        break;
+                    $indexPage->getPantallaAnnul();
+                    break;
                 default:
                     $indexPage->index();
                     break;
